small controller updates

+++ app/controllers/events_controller.rb
@@ -19,24 +19,28 @@ class EventsController < ApplicationController
     end
   end
 
-  #
   def update
     event = Event.find(params[:id])
-    #if current_user created event
-      event.update
-      render json: event
-    #end
+    if event.user_id == current_user.id
+      event.update(event_params)
+      if event.valid?
+        render json: event
+      else
+        render json: event.errors
+      end
+    else
+      render json: {status: 401, error: "You are not the owner of this event"}
+    end
   end
   
-  #
   def destroy
     event = Event.find(params[:id])
-    #if current_user created event
+    if event.user_id == current_user.id
       event.destroy
       render json: event
-    #else
-    
-    #end
+    else
+      render json: {status: 401, error: "You are not the owner of this event"}
+    end
   end
 
   ########
@@ -47,10 +51,11 @@ class EventsController < ApplicationController
     params.require(:event).permit(:image, :category, :about, :event_name)
   end
 
-  def authorize_item
-    unless event.user == current_user 
-      redirect_to items_path, error: 'You are not authorized'
-    end
-  end
+  ###Not important yet, might be later.
+  # def authorize_item
+  #   unless event.user == current_user 
+  #     redirect_to items_path, error: 'You are not authorized'
+  #   end
+  # end
 
 end
